// src/components/atoms/submitButton/types.ts
export interface StyledButtonProps {
  $backgroundColor?: string
  $border?: string
  $borderRadius?: string
  $display?: string
  $fontFamily?: string
  $fontSize?: string
  $fontWeight?: string
  $hoverBackgroundColor?: string
  $hoverBorder?: string
  $hoverTextColor?: string
  $letterSpacing?: string
  $margin?: string
  $maxWidth?: string
  $outerMargin?: string
  $outerPadding?: string
  $padding?: string
  $position?: string
  $textAlign?: string
  $textColor?: string
  $textDecoration?: string
  $textTransform?: string
  $width?: string
}